main.ipynb: updating meta table loader
We changed the dataset to no longer store a pandas DataFrame as this relies on pickle and causes interoperability issues. Instead, the table is now stored as a CSV table bytearray (utf-8). We've also replaced the ID lists with a string of "start:end" which we now have to unwrap upon loading.